Allow images to be rendered from transformer remark
