+++ packages/blog/gatsby-node.js
@@ -33,7 +33,7 @@ exports.createPages = async ({ graphql, actions }) => {
   result.data.allMarkdownRemark.edges.forEach(({ node }) => {
     createPage({
       path: node.fields.slug,
-      component: path.resolve(__dirname, './src/templates/blog-post.tsx'),
+      component: path.resolve(__dirname, './src/templates/blog-post/index.tsx'),
       context: {
         // Data passed to context is available
         // in page queries as GraphQL variables.
