+++ packages/blog/package.json
@@ -18,8 +18,10 @@
     "gatsby": "^4.10.0",
     "gatsby-plugin-gtag": "^1.0.13",
     "gatsby-plugin-react-helmet": "^5.10.0",
+    "gatsby-plugin-sharp": "^4.11.1",
     "gatsby-plugin-styled-components": "^5.10.0",
     "gatsby-plugin-typescript": "^4.10.0",
+    "gatsby-remark-images": "^6.11.1",
     "gatsby-source-filesystem": "^4.10.0",
     "gatsby-transformer-remark": "^5.10.0",
     "react": "^17.0.2",
