+++ packages/ui/src/components/typography/A.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect, useRef } from 'react';
 import styled from 'styled-components';
 
 const AStyled = styled.a`
@@ -7,11 +7,28 @@ const AStyled = styled.a`
 `;
 
 interface AProps {
-  children: React.ReactNode;
+  children?: React.ReactNode;
+  style?: string;
 }
 
-const A: React.FC<AProps> = ({style, ...props}) => {
-  return <AStyled target="_blank" rel="noopener noreferrer" {...props} />
+const A: React.FC<AProps> = (props = {}) => {
+  const { style, ...propsWithoutStyle } = props;
+  const ref = useRef<HTMLAnchorElement>(null);
+
+  useEffect(() => {
+    if(ref.current) {
+      ref.current.style.cssText = style
+    }
+
+  }, [ref.current])
+
+  return (
+    <AStyled
+      ref={ref}
+      target="_blank" rel="noopener noreferrer"
+      {...propsWithoutStyle}
+    />
+  );
 }
 
 export default A;
